refactor: tools information into independent component
